+++ lib/cli/index.js
@@ -50,9 +50,9 @@ log.i18nInit().then(() => {
   commander.option('-s, --silent', log.t('cli.output.description.silent'), false)
 }).then(() => {
   // 引入site管理模块
-  require('./site')
+  require('./siteManage')
   // 引入server管理模块
-  require('./server')
+  require('./serverManage')
   // 引入other管理模块
   require('./other')
   // 引入tail模块
